pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "documentor"
version = "0.1.0"
description = "Структурирование документов для RAG с поддержкой LangChain"
readme = "README.md"
authors = ["angrymuskrat <mlhakov2011@gmail.com>"]
license = "MIT"
keywords = ["document", "parsing", "rag", "langchain", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup",
]
packages = [{include = "documentor"}]
include = ["documentor/config/*.yaml"]

[tool.poetry.dependencies]
python = "^3.12"
# LangChain for Document support
langchain-core = ">=0.3.0"
# YAML for configuration
pyyaml = ">=6.0"
# DOCX processing
python-docx = ">=1.1.0"
# PDF processing
PyMuPDF = ">=1.23.0"
# Image processing
Pillow = ">=10.0.0"
# Data processing
pandas = ">=2.0.0"
# Progress bars
tqdm = ">=4.66.0"
# HTTP requests
requests = ">=2.31.0"
# HTML parsing (for table parsing from OCR)
beautifulsoup4 = ">=4.12.0"
# OpenAI API (for Dots.OCR integration)
openai = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-cov = ">=4.1.0"

[tool.poetry.urls]
Homepage = "https://github.com/yourusername/documentor"
Documentation = "https://github.com/yourusername/documentor#readme"
Repository = "https://github.com/yourusername/documentor"
Issues = "https://github.com/yourusername/documentor/issues"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--cov=documentor",
    "--cov-report=term-missing",
    "--cov-report=html",
]
